+++ ctfbot-rust/src/services/mod.rs
@@ -1,0 +1,2 @@
+pub mod ctftime_service;
+pub mod alpacahack_service;
